more comprehensive tracking of local modules
Module support in u-root works by partitioning modules into local and
external. Local modules have an associated rewrite rule in the go mod
file and their sources are copied into the temporary u-root build area.
In the current code this becomes problematic in the following situation.
Consider the following dependency chains.

rygoo/cmds/my-uroot-prog -> alice/pkg/audio -> bob/pkg/stereo
rygoo/cmds/my-uroot-prog -> bob/pkg/mono

In this situation, I will get the code for /bob/pkg/mono copied, but
not the code for bob/pkg/stereo and u-root will fail to compile my
module. This is because the locality of modules does not take into
account any modules that were considered to be local due to other imports.
This patch builds a set of local modules beforehand and decides to copy
code if the base of the module is in this pre-computed set rather than
only considering the current import context.

Signed-off-by: Ryan Goodfellow <rgoodfel@isi.edu>